+++ measures/CHANGELOG.md
@@ -1,8 +1,13 @@
 # Changelog for `measures`
 
-## 0.1.0.3
+## 0.2.0.0
 
-*
+* Removed `Eq` superclass from `Measure`. I now see some types can not and/or
+  should not provide it.
+
+* Added a `Bounded` instance for `Measure` instances.
+
+* Added the `ByteSize` measure.
 
 ## 0.1.0.2